requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]

# AWS S3 integration
boto3
aioboto3

# Database
psycopg2-binary
asyncpg  # For async database connections

# Pinecone for vector databases
pinecone-client

# OpenAI GPT API and LLM utilities
openai
transformers

# Fine-tuning and dataset handling
peft
datasets
bitsandbytes  # For efficient memory usage

# Environment variable management
python-dotenv

# Logging and Observability
loguru
opentelemetry-api
opentelemetry-sdk

# Async HTTP requests
httpx
aiohttp

# Testing and Mocking
pytest
pytest-asyncio
pytest-mock

# Authentication and Security
pyjwt

# Redis Caching and Rate Limiting
redis>=5.0.0

# Token Handling for OpenAI
tiktoken

# Development Tools
black
isort
flake8

# WebRTC and Real-time Communication
aiortc==1.5.0  # WebRTC implementation
websockets==11.0.3  # WebSocket support
av==10.0.0  # Required for aiortc media handling
cryptography==41.0.5  # Required for secure WebRTC connections
pyee==11.0.1  # Event emitter for WebRTC events
aiodns==3.1.1  # Async DNS resolution
python-multipart==0.0.6  # For handling file uploads
google-crc32c==1.5.0  # For data integrity checks

# Optional WebRTC Media Extensions
numpy>=1.24.0  # For media processing
opencv-python-headless>=4.8.0  # For video processing (if needed)
sounddevice>=0.4.6  # For audio processing (if needed)

pandas>=2.0.0
SQLAlchemy>=2.0.0
python-magic>=0.4.27
textract>=1.6.5
beautifulsoup4>=4.12.0
aiofiles>=23.1.0

qdrant-client>=1.6.0

langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
PyPDF2>=3.0.0

chromadb>=0.4.22

google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
O365>=2.0.0
tenacity>=8.0.0
pytz>=2024.1